app/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.10
requests==2.31.0
pytest==8.3.3
anyio
